+++ packages/fsapp/src/beta-app/shell.web/index.ts
@@ -1,0 +1,9 @@
+export { WebShellContext, WebShellProvider, useWebShell } from './shell.context';
+export { makeDrawer } from './make-drawer';
+export type {
+  WebShell,
+  DrawerComponentProps,
+  DrawerComponentType,
+  DrawerOptions,
+  ShellConfig
+} from './types';